+++ packages/components/src/components/accordion/accordion.browser.e2e.tsx
@@ -1,6 +1,25 @@
 import { describe } from "vitest";
 import { mount } from "@arcgis/lumina-compiler/testing";
-import { defaults, reflects, hidden, renders } from "../../tests/commonTests/browser";
+import { h } from "@arcgis/lumina";
+import { accessible, defaults, reflects, hidden, renders } from "../../tests/commonTests/browser";
+
+describe("accessible", () => {
+  accessible(() =>
+    mount(
+      <calcite-accordion>
+        <calcite-accordion-item heading="Accordion Title 1">
+          Accordion Item Content
+        </calcite-accordion-item>
+        <calcite-accordion-item description="A description" expanded heading="Accordion Title 2">
+          Accordion Item Content
+        </calcite-accordion-item>
+        <calcite-accordion-item heading="Accordion Title 3">
+          Accordion Item Content
+        </calcite-accordion-item>
+      </calcite-accordion>,
+    ),
+  );
+});
 
 describe("defaults", () => {
   defaults(