+++ packages/components/src/components/block-section/block-section.browser.e2e.tsx
@@ -8,8 +8,55 @@ import {
   hidden,
   renders,
   t9n,
+  accessible,
 } from "../../tests/commonTests/browser";
 import { CSS } from "./resources";
+
+describe("accessible", () => {
+  describe("default", () => {
+    describe("expanded", () => {
+      accessible(() =>
+        mount(
+          <calcite-block-section expanded text="text">
+            <div>some content</div>
+          </calcite-block-section>,
+        ),
+      );
+    });
+
+    describe("collapsed", () => {
+      accessible(() =>
+        mount(
+          <calcite-block-section text="text">
+            <div>some content</div>
+          </calcite-block-section>,
+        ),
+      );
+    });
+  });
+
+  describe("toggle-display = 'switch'", () => {
+    describe("expanded", () => {
+      accessible(() =>
+        mount(
+          <calcite-block-section expanded text="text" toggle-display="switch">
+            <div>some content</div>
+          </calcite-block-section>,
+        ),
+      );
+    });
+
+    describe("collapsed", () => {
+      accessible(() =>
+        mount(
+          <calcite-block-section text="text" toggle-display="switch">
+            <div>some content</div>
+          </calcite-block-section>,
+        ),
+      );
+    });
+  });
+});
 
 describe("defaults", () => {
   defaults(