+++ packages/components/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.0-next.15](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.2.0-next.14...@esri/calcite-components@5.2.0-next.15) (2026-07-01)
+
+### Features
+
+- **table:** enable sticky header ([#13906](https://github.com/Esri/calcite-design-system/issues/13906)), closes [#9799](https://github.com/Esri/calcite-design-system/issues/9799)
+
+## [5.2.0-next.14](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.2.0-next.13...@esri/calcite-components@5.2.0-next.14) (2026-07-01)
+
+### Features
+
+- **combobox, input-time-zone, segmented-control:** add corner-radius tokens ([#14402](https://github.com/Esri/calcite-design-system/issues/14402)), closes [#14353](https://github.com/Esri/calcite-design-system/issues/14353)
+
+## [5.2.0-next.13](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.2.0-next.12...@esri/calcite-components@5.2.0-next.13) (2026-07-01)
+
+**Note:** Version bump only for package @esri/calcite-components
+
+## [5.2.0-next.12](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.2.0-next.11...@esri/calcite-components@5.2.0-next.12) (2026-06-30)
+
+### Features
+
+- **panel, flow-item:** add dialog role and focus trap support when closable ([#14685](https://github.com/Esri/calcite-design-system/issues/14685)), closes [#6951](https://github.com/Esri/calcite-design-system/issues/6951)
+
+## [5.2.0-next.11](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.2.0-next.10...@esri/calcite-components@5.2.0-next.11) (2026-06-30)
+
+### Bug Fixes
+
+- **components:** harden sortable touch drag ghost lifecycle ([#14717](https://github.com/Esri/calcite-design-system/issues/14717)), closes [#14678](https://github.com/Esri/calcite-design-system/issues/14678)
+
+## [5.2.0-next.10](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.2.0-next.9...@esri/calcite-components@5.2.0-next.10) (2026-06-30)
+
+### Bug Fixes
+
+- **icon:** fix resolution of number-suffixed icons ([#14711](https://github.com/Esri/calcite-design-system/issues/14711))
+
+## [5.2.0-next.9](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.2.0-next.8...@esri/calcite-components@5.2.0-next.9) (2026-06-29)
+
+### Bug Fixes
+
+- **rating:** fix block size to not encroach on neighboring elements ([#14709](https://github.com/Esri/calcite-design-system/issues/14709)), closes [#14708](https://github.com/Esri/calcite-design-system/issues/14708)
+
 ## [5.2.0-next.8](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components@5.2.0-next.7...@esri/calcite-components@5.2.0-next.8) (2026-06-29)
 
 **Note:** Version bump only for package @esri/calcite-components